+++ customer_portal/src/epilot/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.1.1'
-    gen_version: str = '2.169.3'
-    user_agent: str = 'speakeasy-sdk/python 2.1.1 2.169.3 1.0.0 epilot-customer-portal'
+    sdk_version: str = '2.1.2'
+    gen_version: str = '2.171.0'
+    user_agent: str = 'speakeasy-sdk/python 2.1.2 2.171.0 1.0.0 epilot-customer-portal'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: