+++ customer_portal/RELEASES.md
@@ -661,3 +661,11 @@ Based on:
 - Speakeasy CLI 1.105.0 (2.169.3) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v2.1.1] customer_portal
+
+## 2023-10-26 00:29:32
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/customer-portal.yaml
+- Speakeasy CLI 1.107.0 (2.171.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.1.2] customer_portal